Update. Разделение ответственности. Выделены функции для получения локации размещения и логики расстановки прямоугольников
