+++ TagsCloudContainer.TagsCloudVisualization/Logic/Strategies/Interfaces/IRectanglePlacementStrategy.cs
@@ -1,0 +1,9 @@
+﻿using System.Drawing;
+
+namespace TagsCloudContainer.TagsCloudVisualization.Logic.Strategies.Interfaces;
+
+public interface IRectanglePlacementStrategy
+{
+    public void SetCenterPoint(Point center);
+    public Point GetNextRectangleLocation(Size rectangleSize);
+}
